requirements.txt
asgiref==3.7.2
click==8.1.3
colorama==0.4.6
distlib==0.3.6
Django==4.2.3
djangorestframework==3.14.0
dnspython==2.3.0
filelock==3.12.1
platformdirs==3.5.3
pymongo==4.4.0
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
pyqt5-tools==5.15.9.3.3
PyQt5Designer==5.14.1
python-dotenv==1.0.0
pytz==2023.3
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
virtualenv==20.23.0
django-extensions==3.2.3
